requirements.txt
requests==2.26.0
newspaper3k==0.2.8
beautifulsoup4==4.10.0
html2text==2020.1.16
gne
kuser_agent